Rename message  🐿 v2.8.0
