+++ main.scss
@@ -1,4 +1,4 @@
 @import './src/components/invite-colleagues/main';
 @import './src/components/desktop-app-banner/main';
 @import './src/components/anon-subscribe-now-teal/main';
-@import './src/components/registration-notice/main';
+@import './src/components/team-trial/main';
