+++ manifest.js
@@ -28,9 +28,9 @@ module.exports = {
 		partial: 'top/payment-failure',
 		messageId: 'paymentFailure'
 	},
-	registrationNotice: {
-		partial: 'top/registration-notice',
-		messageId: 'registrationNotice'
+	teamTrial: {
+		partial: 'top/team-trial',
+		messageId: 'teamTrial'
 	},
 	anonSubscribeNow: {
 		partial: 'top/anon-subscribe-now-teal',
