+++ src/components/team-trial/main.scss
